Update readme, do some cleanup
